+++ spec/features/profile_spec.rb
@@ -25,14 +25,14 @@ RSpec.feature "Profile page" do
 
     scenario "I can edit my profile interests" do
       c1, c2 = create_list(:category, 2)
-      ra1, ra2 =  create_list(:research_area, 2)
+      sd1, sd2 =  create_list(:scientific_domain, 2)
 
       visit profile_path
 
       click_link("Edit", match: :first)
 
       select c1.name, from: "Categories of interests"
-      select ra2.name, from: "Research areas of interests"
+      select sd2.name, from: "Scientific domains of interests"
 
       check "New service in Category of interest"
 
@@ -42,28 +42,28 @@ RSpec.feature "Profile page" do
         expect(page).to have_text(c1.name)
         expect(page).to_not have_text(c2.name)
 
-        expect(page).to_not have_text(ra1.name)
-        expect(page).to have_text(ra2.name)
+        expect(page).to_not have_text(sd1.name)
+        expect(page).to have_text(sd2.name)
       end
 
       expect(page).to have_text("Category of interests")
-      expect(page).to_not have_text("Research area of interests")
+      expect(page).to_not have_text("Scientific domain of interests")
     end
 
     scenario "I can delete my profile information" do
       c = create(:category)
-      ra = create(:research_area)
-      user.update(categories: [c], research_areas: [ra],
-                  categories_updates: true, research_areas_updates: true)
+      sd = create(:scientific_domain)
+      user.update(categories: [c], scientific_domains: [sd],
+                  categories_updates: true, scientific_domains_updates: true)
 
       visit profile_path
 
       within "#profile-information" do
         expect(page).to have_text(c.name)
-        expect(page).to have_text(ra.name)
+        expect(page).to have_text(sd.name)
 
         expect(page).to have_text("Category of interests")
-        expect(page).to have_text("Research area of interests")
+        expect(page).to have_text("Scientific domain of interests")
       end
 
       click_link("Edit", match: :first)
@@ -74,10 +74,10 @@ RSpec.feature "Profile page" do
 
       within "#profile-information" do
         expect(page).to_not have_text(c.name)
-        expect(page).to_not have_text(ra.name)
+        expect(page).to_not have_text(sd.name)
 
         expect(page).to_not have_text("Category of interests")
-        expect(page).to_not have_text("Research area of interests")
+        expect(page).to_not have_text("Scientific domain of interests")
       end
     end
   end